+++ docs/changelog.md
@@ -7,3 +7,8 @@ ridheshcybe
 authored
 Verified
 - Porting to react (Added index route) - 13/10/2025
+
+ridheshcybe
+authored
+Verified
+- Added login and signup routes (Implemeting later) - 13/10/2025